[alignmentManager] waits for the actionRecognizer to output a prediction

+++ modules/motionAnalyzer/include/Metric.h
@@ -28,7 +28,7 @@ protected:
     std::string motion_type;
     std::string tag_joint;
     std::string tag_plane;
-    double duration;
+    int duration;
     yarp::sig::Vector ref_dir;
     double min;
     double max;
@@ -51,7 +51,7 @@ public:
     void print();
     void initialize(const std::string &name_, const std::string &motion_type_, const std::string &tag_joint_,
                     const yarp::sig::Vector &ref_dir_, const std::string &tag_plane_, const double &min_, const double &max_,
-                    const double &duration_, const yarp::sig::Vector &camerapos_,
+                    const int &duration_, const yarp::sig::Vector &camerapos_,
                     const yarp::sig::Vector &focalpoint_, const std::vector<std::string> &relaxed_joints_,
                     const yarp::sig::Vector &dtw_thresh_, const yarp::sig::Vector &mean_thresh_,
                     const yarp::sig::Vector &sx_thresh_, const yarp::sig::Vector &sy_thresh_, const yarp::sig::Vector &sz_thresh_,
@@ -60,7 +60,7 @@ public:
     yarp::sig::Vector getRefDir() const { return ref_dir; }
     double getMax() const { return max; }
     double getMin() const { return min; }
-    double getDuration() const { return duration; }
+    int getDuration() const { return duration; }
     std::string getName() const { return name; }
     std::string getTagPlane() const { return tag_plane; }
     std::string getTagJoint() const { return tag_joint; }
