+++ modules/motionAnalyzer/src/Metric.cpp
@@ -27,7 +27,7 @@ Metric::~Metric()
 }
 
 void Metric::initialize(const string &name_, const string &motion_type_, const string &tag_joint_, const Vector &ref_dir_,
-                        const string &tag_plane_, const double &min_, const double &max_, const double &duration_,
+                        const string &tag_plane_, const double &min_, const double &max_, const int &duration_,
                         const Vector &camerapos_, const Vector &focalpoint_, const vector<string> &relaxed_joints_,
                         const Vector &dtw_thresh_, const Vector &mean_thresh_, const Vector &sx_thresh_,
                         const Vector &sy_thresh_, const Vector &sz_thresh_, const Vector &f_static_,
